Support full mesh and partial mesh Auto VPN topology.
The existing code assumes Hub & Spoke Auto VPN topology with vMX as the Hub.
Full and partial mesh are not assumed.
For example, the Auto VPN topology where the Hubs exist in the data center and AWS. In this case, the script does not get route to the data center.

Modify the route learning logic to support full and partial mesh.